Wind 1.3.2
04/10/2020, Bing Li

When running the streaming framework, I noticed that the thread count is always increased even though the publishers post data in a low frequency.

I found bugs exist in the NotificationDispatcher/NotificationQueue, RequestDispatcher/RequestQueue, and so forth.

The bugs are caused by the judging the states of threads (NotificationQueue/RequestQueue) and thread pools (NotificationDispatcher/RequestDispatcher) upon timeout.

Java condition await() returns false when the predefined timeout is not reached. It returns true when the predefined timeout is reached.

I did not return the boolean value in the previous design. That causes the wrong judgment about the states.

So the thread pools (NotificationDispatcher/RequestDispatcher) are always executed by Executor of Java.

And the threads (NotificationQueue/RequestQueue) are always removed to cause new threads creations.

They are not reused properly.

Those code are the fundamentals of GreatFree such that it is required to keep them high quality.

One potential problems, why the superfluous threads are not killed. I need to check the problem.

The current lines of code are 200721.

The number of files is 2966.