Version 1.3.8.0
09/14/2020, Bing Li

One important update to the version is the synchronization of the children cluster through the root.

In some cases, it is necessary to synchronize among the children, e.g. the storage children need to keep their paths consistent if they are located on the same machine.

I have ever used file locking to do that.

But it is not good.

Through the root, the approach is reasonable and convenient.

The current lines of code are 205865.

The number of files is 3083.