Version 1.3.9.2
08/30/2021, Bing Li

Two simple methods are added to Rand to generate characters in String.

The current lines of code are 205931.

The number of files is 3084.

================================================

08/28/2021, Bing Li

ServerProfile is not initialized properly.

When distributed nodes are deployed on different computing devices, the registry server IP needs to be retained in ServerProfile.

But since I did not test the case, the registry server IP is always 127.0.0.1.

The bug is fixed.

The current lines of code are 205916.

The number of files is 3084.