Wind 1.4.1.0
11/08/2021, Bing Li

A critical bug is fixed!

It exists in the DP, the NotificationDispatcher and the RequestDispatcher.

As a thread pool for the messaging-based threads, one problem is misunderstood.

Condition.await() returns false when the predefined timeout is passed completely.

Rather, if the waiting is interrupted or signaled, it returns true.

I made a mistake on that.

As such, threads are never killed even though they are idle since they are not correctly judged to be idle.

Another problem is that the NotificationDispatcher and the RequestDispatcher cannot stop themselves by themselves.

As the important mechanisms to manage threads, they are shutdown only if programmers decide to do so in one specific scenario.

The current lines of code are 207132.

The number of files is 3107.