Version 1.4.8.8
I am writing a book about GreatFree/Wind. It must help developers understand the philosophy of GreatFree. Rich sample code is available in the book. This book must be the first one in China. Even for the world, the technology in the book must make important contributions to the domain of distributed programming over the Internet. The code of GreatFree is updated gradually. Now the latest version is 1.4.8.8. You are welcome to program with GreatFree.